Update local json file format
The json file will now only include an AssetMap for each provider and will be identified by the provider's name. The file will eventually be used to initialize the db and will not be referenced otherwise. The existing files that rely on this json file have been updated so they are in a working state as of this commit but the mechanism of accessing unique IDs will change in an upcoming commit.