Remove default currency codes from incoming uid maps
If a provider supports a different asset with the same code as our default, we risked querying for that other asset instead. This change means we'll only add uids for currency codes we're not already looking for. The one limitation is if we need to add or change a default UID we'll need to commit it and restart the process so it takes effect.