Use case-appropriate values for assetMap and DB saved values
This resolves issues with mixed-case currency codes like is used for XRP tokens. The assetMap should use the actual case-sensitive string like a client would request. The value that is added to the rates object, which is saved to the DB, needs to be all caps (except the fiat portion).  Unfortunately, since each provider handles currency code casing uniquely it kind of has to be handled here and not in a more general location which is why this is kind of ugly.